refactor :  모듈화되지 않은 컴포넌트 정리
