+++ src/components/CurrentVote/index.ts
@@ -1,0 +1,3 @@
+export { default as CurrentVoteBtn } from './CurrentVoteBtn';
+export { default as CurrentVoteInfo } from './CurrentVoteInfo';
+export { default as CurrentVoteSlider } from './CurrentVoteSlider';
