+++ src/components/Landing/maker/index.ts
@@ -1,2 +1,4 @@
+export { default as LandingCurrentVote } from './LandingCurrentVote';
+export { default as LandingHeader } from './LandingHeader';
 export { default as LandingVoteCard } from './LandingVoteCard';
 export { default as LandingVoteList } from './LandingVoteList';
