+++ src/components/common/index.ts
@@ -1,3 +1,4 @@
 export { default as Error } from './Error';
 export { default as Loading } from './Loading';
+export { default as Modal } from './Modal';
 export { default as StickerAttachImg } from './StickerAttachImg';
